ffi: Move error type into its own module
